pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "beatsynchr"
version = "0.1.0"
description = "A Python library for syncing images and videos with music beats"
readme = "README.md"
authors = [
    { name = "Jakub Nurkiewicz", email= "jakub.nurkiewicz.2003@gmail.com" }
]
keywords = ["images processing", "video processing", "music beats"]
classifiers = [
    "Programming Language :: Python :: 3.12"
]
dependencies = [
    "librosa>=0.11.0",
    "numpy>=2.0.0",
    "moviepy>=2.0.0",
    "Pillow>=11.0.0",
    "pillow-heif>=1.0.0"
]
requires-python = ">=3.12"